Avoid the term whitelist
This term was replaced with allowedList. See e.g. https://tools.ietf.org/id/draft-knodel-terminology-00.html
for a rationale.
