+++ CHANGES
@@ -7,6 +7,8 @@ jsoup changelog
 
   * Improvement: added support for loading and parsing gzipped HTML files in Jsoup.parse(File in, charset, baseUri).
 
+  * Change: use Allowlist instead of Whitelist in HTML sanitizer
+
 *** Release 1.13.1 [2020-Feb-29]
   * Improvement: added Element#closest(selector), which walks up the tree to find the nearest element matching the
     selector.
@@ -445,7 +447,7 @@ jsoup changelog
  * Added the new selector :containsData(), to find elements that hold data, like script and style tags.
 
  * Changed Jsoup.isValid(bodyHtml) to validate that the input contains only body HTML that is safe according to the
-   whitelist, and does not include HTML errors. And in the Jsoup.Cleaner.isValid(Document) method, make sure the doc
+   allowlist, and does not include HTML errors. And in the Jsoup.Cleaner.isValid(Document) method, make sure the doc
    only includes body HTML.
    <https://github.com/jhy/jsoup/issues/245>
    <https://github.com/jhy/jsoup/issues/632>
@@ -721,7 +723,7 @@ jsoup changelog
  * If pretty-print is disabled, don't trim outer whitespace in Element.html()
    <https://github.com/jhy/jsoup/issues/368>
 
- * In the HTML Cleaner, allow span tags in the basic whitelist, and span and div tags in the relaxed whitelist.
+ * In the HTML Cleaner, allow span tags in the basic allowlist, and span and div tags in the relaxed allowlist.
 
   * Added Element.cssSelector(), which returns a unique CSS selector/path for an element.
     <https://github.com/jhy/jsoup/pull/459>
@@ -912,7 +914,7 @@ jsoup changelog
  * Fixed NPE when HTML fragment parsing a <style> tag
    <https://github.com/jhy/jsoup/issues/189>
 
- * Fixed issue with :all pseudo-tag in HTML sanitizer when cleaning tags previously defined in whitelist
+ * Fixed issue with :all pseudo-tag in HTML sanitizer when cleaning tags previously defined in allowlist
    <https://github.com/jhy/jsoup/issues/156>
 
  * Fixed NPE in Parser.parseFragment() when context parameter is null.
@@ -1242,7 +1244,7 @@ jsoup changelog
     prepend(String), append(String); bulk methods for corresponding
     methods in Element.
 
- * New feature: Jsoup.isValid(html, whitelist) method for user input
+ * New feature: Jsoup.isValid(html, allowlist) method for user input
     form validation.
  
  * Improved Elements.attr(String) to find first matching element
